# api/openapi.yaml
openapi: 3.0.3
info:
  title: x-ui
  version: "1.0.0"
  description: |
    Control-plane API for a tunnel panel: operator accounts, subscribers
    (clients), listeners (inbounds) with the TLS each terminates with, reported
    traffic, and the audit log.

    Every endpoint answers with the same envelope:

        { "success": true,  "msg": "", "obj": <payload> }
        { "success": false, "msg": "<reason>", "obj": null }

    `success` mirrors the status code; it is there so a client can branch on
    one field rather than on a status line it may not see through a proxy.

    The same endpoints are mounted twice, under two different authentication
    schemes. `{basePath}api` takes a session cookie and is what a browser uses;
    `{basePath}apiv2` takes a `Token` header and is what a script or a node
    uses. Only the cookie API enforces same-origin, because a cross-site page
    cannot set a custom header without a CORS preflight that is never answered.

    Only GET, POST and DELETE are used. An update is a POST to the record's own
    path rather than a PUT: a panel is reached through whatever sits in front of
    it — a corporate proxy, an embedded webview, a captive network's filter —
    and the three are the ones those are relied on to forward.
servers:
  - url: "http://localhost:8000"
    description: Panel — operators, authenticated
  - url: "http://localhost:8443"
    description: Subscriptions — subscribers, unauthenticated, its own listener
...
paths:
...
  /api/core/restart:
    post:
      tags: [system]
      summary: Queue a restart of the local sing-box service
      description: |
        Requires the root-managed systemd installation, a managed local agent,
        native statistics and the updated reload helper. Validates the actual
        service configuration under the agent's application lock before restarting.
        Existing connections close; the panel remains available. Success requires
        a new PID and a healthy native statistics API. A duplicate request while
        active returns the existing task. Saved Panel settings are not applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "202":
          description: Restart queued or existing active task
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/CoreRestartJob" }
        "400":
          $ref: "#/components/responses/Invalid"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/core/restart/{id}:
    get:
      tags: [system]
      summary: Follow a persisted local core restart task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, pattern: "^[a-f0-9]{32}$" }
      responses:
        "200":
          description: Current progress or terminal result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/CoreRestartJob" }
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          description: Task does not exist
  /api/core/logs:
    get:
      tags: [system]
      summary: Read the last 80 local sing-box journal entries
      responses:
        "200":
          description: Journal output capped at 64 KiB
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj:
                        type: object
                        properties:
                          lines:
                            type: array
                            items: { type: string }
        "400":
          $ref: "#/components/responses/Invalid"
...
  /api/keypairs:
    get:
      tags: [system]
      summary: Generate key material
      description: |
        Generated, returned once, never stored, and answered with
        `Cache-Control: no-store`. It exists because the alternative to
        generating a private key here is generating it on a web page somebody
        else runs.

        ECH keys are not generated: they need the core's own implementation, and
        producing something that merely looks like one would be worse.
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum: [reality, wireguard, tls, openvpn]
        - name: options
          in: query
          schema: { type: string }
          description: |
            For `tls`, the server name or address the certificate is for. For
            `wireguard`, an existing private key, in which case only the
            matching public key is returned and the private key is not echoed
            back.
      responses:
        "200":
          description: The key material
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/Keypair" }
        "400":
          $ref: "#/components/responses/Invalid"
  /api/cert-probe:
    post:
      tags: [system]
      summary: Look at the certificate a host presents
      description: |
        Verification is deliberately skipped — refusing to look at a certificate
        because it does not verify is exactly backwards for a diagnostic. The
        result is described, never trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, description: A host name, not a URL. }
                port: { type: string, default: "443" }
      responses:
        "200":
          description: What the host presented
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/CertificateProbe" }
        "400":
          $ref: "#/components/responses/Invalid"
  /api/backup:
    get:
      tags: [system]
      summary: Export the panel's data
      description: |
        A logical export — a JSON document of every table — not a file copy.
        `pg_dump` remains the right tool for an operational backup; this is the
        one the panel can offer through its own API, and the one that restores
        into a fresh install of a different version.

        It contains every credential the panel holds, so taking one is recorded
        in the audit log and the response carries `Cache-Control: no-store`.
      parameters:
        - name: exclude
          in: query
          schema: { type: string }
          description: Comma-separated tables to leave out, e.g. `stats,changes`.
      responses:
        "200":
          description: The export, as a file
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/backup/restore:
    post:
      tags: [system]
      summary: Replace the panel's data with an export
      description: |
        Runs in a single transaction, so a file that turns out to be unreadable
        halfway through leaves the existing data untouched rather than
        half-replaced. A table the export does not carry is left alone rather
        than emptied.

        The operator accounts come from the backup too, so whoever runs this
        will be signing in with the backup's credentials, not their own.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [backup]
              properties:
                backup: { type: string, format: binary }
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          $ref: "#/components/responses/Invalid"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: x-ui
      description: Set by /api/signin. Used under the `{basePath}api` prefix.
    token:
      type: apiKey
      in: header
      name: Token
      description: Minted by /api/tokens. Used under the `{basePath}apiv2` prefix.
...
  responses:
    Invalid:
      description: The request has to be fixed before it can succeed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - type: object
                properties:
                  obj: { $ref: "#/components/schemas/ValidationErrors" }
...
    Conflict:
      description: Collides with something already stored
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        msg: { type: string }
        obj: { nullable: true }
    ValidationErrors:
      type: object
      properties:
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
...
    CoreRestartJob:
      type: object
      required: [id, state, actor, requestedAt, beforePid, afterPid]
      properties:
        id: { type: string, pattern: "^[a-f0-9]{32}$" }
        state: { type: string, enum: [queued, checking, restarting, verifying, succeeded, failed] }
        actor: { type: string }
        requestedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        beforePid: { type: integer, minimum: 0 }
        afterPid: { type: integer, minimum: 0 }
        error: { type: string, description: Public restart failure message. }
...
    Keypair:
      type: object
      properties:
        kind: { type: string, enum: [reality, wireguard, tls, openvpn] }
        privateKey: { type: string, description: The half that stays on the server. Absent when an existing one was supplied. }
        publicKey: { type: string, description: The half that goes to subscribers. }
        certificate: { type: string, description: PEM, for the tls kind. }
        secret: { type: string, description: A single shared value, for the openvpn kind. }
        note: { type: string, description: Anything an operator has to know to use it. }
    CertificateProbe:
      type: object
      properties:
        domain: { type: string }
        port: { type: integer }
        subject: { type: string }
        issuer: { type: string }
        dnsNames: { type: array, items: { type: string } }
        notBefore: { type: integer, format: int64 }
        notAfter: { type: integer, format: int64 }
        daysRemaining: { type: integer, format: int64, description: Negative once it has expired. }
        sha256: { type: string, description: The fingerprint a client would pin against. }
        selfSigned: { type: boolean, description: The difference between "needs skip-verify" and "the chain is wrong". }
        tlsVersion: { type: string }
        alpn: { type: string }
...
security:
  - session: []
  - token: []